Function factory for mutual information.

Creates a function to be called in the tidyverse that calculates mutual information between variables of a data set. Intended to be used as part of a group_by(...)

column_info_func(col1, col2)

Arguments

col1

Name of the first column

col2

Name of the second column

Examples

ChickWeight %>% group_by(Diet)  %>% 
    group_map(column_info_func("Chick", "weight")) %>% 
    do.call(rbind, .)
#>      Diet                                    
#> [1,]    1 2.961254 2.197479 2.197479 4.589666
#> [2,]    2 2.302585 1.948594 1.948594 4.410395
#> [3,]    3 2.302585 1.984780 1.984780 4.469686
#> [4,]    4 2.301229 1.958975 1.958975 4.416682